style(msvc): fix C4244/C4267 in OptionsList/EditMenu/CubicSpline/TimingData/MusicWheel/CourseLoaderCRS + 3 residual sites (item 2)
All static_cast, documenting existing intentional narrowing -- no
logic changes. sm_tests unchanged at 5966/226; test_TimingData.cpp
and test_NotesLoaderCRS.cpp re-checked identical.

- OptionsList.cpp (10): unsigned/int locals and a GetScreen(int) call
  fed size_t choice counts/indices; two wrap(int&, size_t) sites; a
  linear-search loop returning its size_t index as int.
- EditMenu.cpp (10): two reverse-loop size()-1 bounds; six one-line
  GetRowSize accessors returning size() as int; a size()-1
  default-selection index.
- CubicSpline.cpp (10): a shared LCSN_EVAL_SOMETHING macro fixed once
  covers 4 call sites; three more lua_createtable/lua_rawgeti/
  lua_rawseti sites in the coefficient get/set helpers; two
  lua_pushnumber(size_t) accessors.
- TimingData.cpp (9, SS5-protected TimingData, characterization-only):
  a 4-way size() sum into an unsigned int segment count; an
  EraseSegment(int, ...) fed a size_t loop index at 2 sites; a
  binary-search size()-1 upper bound; a lua_createtable/lua_rawseti
  block serializing timing segments to Lua.
- MusicWheel.cpp (9): a MusicWheelItemData ctor's int section-count
  param fed size_t at 2 sites; two unsigned snapshot-size locals; a
  std::min<unsigned int> call's second arg; two wrap(int&, size_t)
  sites; a reverse-loop bound; a RandomInt(size_t) call.
- CourseLoaderCRS.cpp (9, SS5-protected .crs, characterization-only):
  four Left/Right pairs fed strlen()/size()-derived lengths parsing
  BEST/WORST/GRADEBEST/GRADEWORST song-choice prefixes; an int
  (StringToInt) assigned into a float field.
- Residual C4267 the broken C424[47] regex had hidden in 3 files
  already marked done earlier this sweep: NoteField.cpp (5: two int
  column-index members fed a size_t loop var, a binary-search
  size()-1 bound, a lua_createtable/lua_rawseti pair pushing
  column-renderer actors); ScreenGameplay.cpp (2: an unsigned int
  queue-size local, a reverse-loop size()-1 bound);
  ScreenOptionsMasterPrefs.cpp (1: a MoveMap(..., unsigned cnt) call
  fed size_t).

+++ src/EditMenu.cpp
@@ -67,7 +67,7 @@ void EditMenu::GetSongsToShowForGroup( const RString &sGroup, std::vector<Song*>
 	case EditMode_Practice:
 	case EditMode_CourseMods:
 	case EditMode_Home:
-		for( int i=vpSongsOut.size()-1; i>=0; i-- )
+		for( int i=static_cast<int>(vpSongsOut.size())-1; i>=0; i-- )
 		{
 			const Song* pSong = vpSongsOut[i];
 			if( !pSong->NormallyDisplayed()  ||  pSong->IsTutorial() )
@@ -89,7 +89,7 @@ void EditMenu::GetGroupsToShow( std::vector<RString> &vsGroupsOut )
 		return;
 
 	SONGMAN->GetSongGroupNames( vsGroupsOut );
-	for( int i = vsGroupsOut.size()-1; i>=0; i-- )
+	for( int i = static_cast<int>(vsGroupsOut.size())-1; i>=0; i-- )
 	{
 		const RString &sGroup = vsGroupsOut[i];
 		std::vector<Song*> vpSongs;
@@ -271,13 +271,13 @@ int EditMenu::GetRowSize( EditMenuRow er ) const
 {
 	switch( er )
 	{
-	case ROW_GROUP:		return m_sGroups.size();
-	case ROW_SONG:			return m_pSongs.size();
-	case ROW_STEPS_TYPE:	return m_StepsTypes.size();
-	case ROW_STEPS:		return m_vpSteps.size();
-	case ROW_SOURCE_STEPS_TYPE:	return m_StepsTypes.size();
-	case ROW_SOURCE_STEPS:	return m_vpSourceSteps.size();
-	case ROW_ACTION:		return m_Actions.size();
+	case ROW_GROUP:		return static_cast<int>(m_sGroups.size());
+	case ROW_SONG:			return static_cast<int>(m_pSongs.size());
+	case ROW_STEPS_TYPE:	return static_cast<int>(m_StepsTypes.size());
+	case ROW_STEPS:		return static_cast<int>(m_vpSteps.size());
+	case ROW_SOURCE_STEPS_TYPE:	return static_cast<int>(m_StepsTypes.size());
+	case ROW_SOURCE_STEPS:	return static_cast<int>(m_vpSourceSteps.size());
+	case ROW_ACTION:		return static_cast<int>(m_Actions.size());
 	default: FAIL_M( ssprintf("Non-existant EditMenuRow %i", er) );
 	}
 }
@@ -472,7 +472,7 @@ void EditMenu::OnRowValueChanged( EditMenuRow row )
 
 					// Try to preserve the user's StepsType selection.
 					if(st == orgSel)
-					m_iSelection[ROW_STEPS_TYPE] = m_StepsTypes.size() - 1;
+					m_iSelection[ROW_STEPS_TYPE] = static_cast<int>(m_StepsTypes.size()) - 1;
 				}
 			}
 		}
